working deposit service, no tests though

+++ src/main/java/com/revature/thevault/service/dto/DepositResponseObject.java
@@ -17,7 +17,7 @@ public class DepositResponseObject {
 	String depositType;
 	String reference;
 	LocalDate dateDeposit;
-	int amount;
+	float amount;
 }
